+++ app/backend/requirements.txt
@@ -1,1 +1,2 @@
 flask
+pytest
